rename geo and photo_save to save_photo/geo move save_photo/geo to class PathGenerator rename PathGeneraor to Storage
