+++ 3750/2/Gemfile
@@ -4,4 +4,6 @@ git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 gem 'activesupport'
 gem 'redis'
+gem 'redis-activesupport'
+gem 'redis-store'
 gem 'telegram-bot'
